# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "MIND Teams Challenge",
        "description": "MIND Teams Challlenge, for implementing docker, swagger and more...",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "components":{
        "securitySchemes":{
            "bearerAuth":{
                "type":"http",
                "scheme":"bearer",
                "bearerFormat":"JWT"
            }
        }
    },
    "security":{
        "bearerAuth":[]
    },
    "host": "localhost:40000",
    "basePath": "/api",
    "paths": {
        "/authenticate": {
            "post": {
                "parameters": [
                    {
                        "name": "objeto",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "example":{"appname":"MINDCHALLENGE","params":{"userName":"yourusername","pass":"yourpassword"}}
                    }
                ],
                "description": "Método para autenticación",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401":{
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/runsp": {
            "post": {
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "example":{"authorization":"sf34x5g..."}
                    },
                    {
                        "name": "objeto",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "example":{"appname":"MINDCHALLENGE","sp":"GetAllUsers","params":[]}
                    }
                ],
                "description": "Ejecución de SPs protegida con tedious mssql",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401":{
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "parameters": [],
                "description": "Endpoint de prueba",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/v2/runsp/": {
            "post": {
                "parameters": [
                    {
                        "name": "objeto",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "example":{"appname":"MINDCHALLENGE","sp":"GetAllUsers","params":[]}
                    }
                ],
                "description": "Ejecución de SPs sin token usando tedious mssql",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/v2/simpleSum/{n1}/{n2}": {
            "get": {
                "parameters": [
                    {
                        "name": "n1",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "n2",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "Suma sencilla en version 2",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/v2/sequelize/": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "sp",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "example":{"sp":"GetAllUsers"}
                    }
                ],
                "description": "Ejecución de SPs sin token usando sequelize",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Challenge",
            "description": "API for MIND Challenge"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "servers": [
        {
            "url": "http://localhost:40000/api/"
        }
    ],
    "apis": [
        "./server.js"
    ]
}